Adding date validation and fix bug converting leap month from lunardate to solardate

+++ src/main/kotlin/com/amlich/VNDate.kt
@@ -26,7 +26,7 @@ class VNDate {
 
     constructor(lunarDate: LunarDate, timeZoneOffset: Int = TimeZoneOffset) {
         this.timeZoneOffset = timeZoneOffset
-        val solarDate = lunar2solar(lunarDate.year, lunarDate.month, lunarDate.day, lunarDate.leap.toInt(), timeZoneOffset)
+        val solarDate = lunar2solar(lunarDate.year, lunarDate.month, lunarDate.day, lunarDate.leap, timeZoneOffset)
         this.solarTime = ZonedDateTime.of(solarDate.year, solarDate.month, solarDate.day, 12, 0, 0, 0, VNTimeZone)
         this.lunarDate = lunarDate
     }
